# Fix #267 - Optimize slow commit performance