--- a/bench/__init__.py
+++ b/bench/__init__.py
@@ -2,17 +2,17 @@
 
 This package provides comprehensive benchmarking capabilities for Oboyu:
 - Speed benchmarks for indexing and search performance
-- Accuracy evaluation for RAG performance 
+- Accuracy evaluation for RAG performance
 - Reranker-specific evaluation
 - Unified benchmark runner
 
 Examples:
     # Run all benchmarks quickly
     python bench/run_benchmarks.py all --quick
-    
-    # Run only speed benchmarks  
+
+    # Run only speed benchmarks
     python bench/run_benchmarks.py speed --datasets small
-    
+
     # Run accuracy evaluation
     python bench/run_benchmarks.py accuracy --datasets synthetic