--- a/bench/accuracy/rag_accuracy/__init__.py
+++ b/bench/accuracy/rag_accuracy/__init__.py
@@ -19,4 +19,3 @@ __all__ = [
     "RerankerEvaluator",
     "ResultsAnalyzer",
 ]
-