--- a/bench/quick_bench.py
+++ b/bench/quick_bench.py
@@ -14,15 +14,16 @@ def run_command(cmd: list[str], description: str) -> bool:
     """Run a command and handle errors."""
     print(f"🔄 {description}")
     try:
-        result = subprocess.run(cmd, check=True, cwd=Path(__file__).parent.parent)
+        subprocess.run(cmd, check=True, cwd=Path(__file__).parent.parent)  # noqa: S603
         print(f"✅ {description} - Complete")
         return True
     except subprocess.CalledProcessError as e:
         print(f"❌ {description} - Failed with exit code {e.returncode}")
         return False
 
+
 def main() -> int:
-    """Main entry point."""
+    """Run quick benchmark operations."""
     parser = argparse.ArgumentParser(description="Quick benchmark operations")
     parser.add_argument(
         "operation",
@@ -39,49 +40,28 @@ def main() -> int:
 
     if args.operation == "setup":
         # Setup test data and queries
-        success &= run_command(
-            ["uv", "run", "python", "bench/speed/generate_test_data.py", "small"],
-            "Generating small test dataset"
-        )
-        success &= run_command(
-            ["uv", "run", "python", "bench/speed/generate_queries.py"],
-            "Generating query datasets"
-        )
+        success &= run_command(["uv", "run", "python", "bench/speed/generate_test_data.py", "small"], "Generating small test dataset")
+        success &= run_command(["uv", "run", "python", "bench/speed/generate_queries.py"], "Generating query datasets")
 
     elif args.operation == "quick":
         # Run quick benchmark suite
-        success &= run_command(
-            ["uv", "run", "python", "bench/run_benchmarks.py", "all", "--quick"],
-            "Running quick benchmark suite"
-        )
+        success &= run_command(["uv", "run", "python", "bench/run_benchmarks.py", "all", "--quick"], "Running quick benchmark suite")
 
     elif args.operation == "speed":
         # Run only speed benchmarks
-        success &= run_command(
-            ["uv", "run", "python", "bench/run_benchmarks.py", "speed", "--datasets", "small"],
-            "Running speed benchmarks"
-        )
+        success &= run_command(["uv", "run", "python", "bench/run_benchmarks.py", "speed", "--datasets", "small"], "Running speed benchmarks")
 
     elif args.operation == "accuracy":
         # Run only accuracy benchmarks
-        success &= run_command(
-            ["uv", "run", "python", "bench/run_benchmarks.py", "accuracy", "--datasets", "synthetic"],
-            "Running accuracy benchmarks"
-        )
+        success &= run_command(["uv", "run", "python", "bench/run_benchmarks.py", "accuracy", "--datasets", "synthetic"], "Running accuracy benchmarks")
 
     elif args.operation == "all":
         # Run comprehensive benchmarks
-        success &= run_command(
-            ["uv", "run", "python", "bench/run_benchmarks.py", "all", "--comprehensive"],
-            "Running comprehensive benchmark suite"
-        )
+        success &= run_command(["uv", "run", "python", "bench/run_benchmarks.py", "all", "--comprehensive"], "Running comprehensive benchmark suite")
 
     elif args.operation == "analyze":
         # Analyze latest results
-        success &= run_command(
-            ["uv", "run", "python", "bench/speed/analyze.py", "--latest", "3"],
-            "Analyzing latest speed benchmark results"
-        )
+        success &= run_command(["uv", "run", "python", "bench/speed/analyze.py", "--latest", "3"], "Analyzing latest speed benchmark results")
 
     if success:
         print("\n🎉 All operations completed successfully!")
@@ -89,6 +69,7 @@ def main() -> int:
     else:
         print("\n💥 Some operations failed. Check the output above.")
         return 1
+
 
 if __name__ == "__main__":
     sys.exit(main())