--- a/docs/sidebars.ts
+++ b/docs/sidebars.ts
@@ -123,6 +123,7 @@ const sidebars: SidebarsConfig = {
             'for-developers/japanese',
             'for-developers/e2e_display_testing',
             'for-developers/pre-commit-optimization',
+            'for-developers/ci-cd-optimization',
             'for-developers/release-process',
           ],
         },