--- a/tests/cli/test_index_manage.py
+++ b/tests/cli/test_index_manage.py
@@ -1,14 +1,9 @@
 """Tests for the Oboyu CLI index manage commands."""
 
-import tempfile
-from pathlib import Path
 from unittest.mock import patch
 
 import pytest
 from typer.testing import CliRunner
-
-from oboyu.cli.manage import app as manage_app
-from oboyu.common.config import ConfigManager
 
 runner = CliRunner()