--- a/tests/cli/utils/test_utils.py
+++ b/tests/cli/utils/test_utils.py
@@ -1,7 +1,5 @@
 """Tests for the Oboyu CLI utilities."""
 
-import pytest
-
 from oboyu.cli.utils import contains_japanese, format_snippet