… .github/workflows/pr-validation.yml
… .github/workflows/test-pypi.yml
… .pre-commit-config.yaml
… CLAUDE.md
… bench/accuracy/rag_accuracy/dataset_manager.py
… bench/config.py
… bench/core/benchmark_base.py
… bench/core/metrics.py
… bench/reranker/benchmark_reranking.py
… bench/speed/analyze.py
… bench/speed/benchmark_indexing.py
… bench/speed/benchmark_search.py
… bench/speed/generate_queries.py
… bench/speed/generate_test_data.py
… bench/speed/reporter.py
… bench/speed/results.py
… bench/speed/run_speed_benchmark.py
… bench/speed/runner.py
… bench/utils.py
… docs/for-developers/ci-cd-optimization.md
… docs/for-developers/pre-commit-optimization.md
… docs/pre-commit-optimization.md
… pyproject.toml
… tests/__init__.py
… tests/cli/__init__.py
… tests/cli/commands/__init__.py
… tests/cli/commands/test_index.py
… tests/cli/commands/test_mcp.py
… tests/cli/commands/test_query.py
… tests/cli/test_clear.py